+++ FluoRender/FluoRender/Format/base_reader.cpp
@@ -274,7 +274,7 @@ int BaseReader::GetOffset() { return m_cur_batch; }
 int BaseReader::get_number(string &str, int64_t pos)
 {
 	string num_str;
-	for (int64_t i = pos; i < str.length(); ++i)
+	for (int64_t i = pos; i < static_cast<int64_t>(str.length()); ++i)
 	{
 		if (isdigit(str[i]))
 			num_str.push_back(str[i]);
@@ -290,7 +290,7 @@ int BaseReader::get_number(string &str, int64_t pos)
 double BaseReader::get_double(string& str, int64_t pos)
 {
 	string num_str;
-	for (int64_t i = pos; i < str.length(); ++i)
+	for (int64_t i = pos; i < static_cast<int64_t>(str.length()); ++i)
 	{
 		if (isdigit(str[i]) || str[i] == '.' ||
 			(i == pos && str[i] == '-'))
